requirements.txt
attrs==22.2.0
autopep8==2.0.2
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
chardet==5.1.0
charset-normalizer==2.0.10
colorama==0.4.6
decorator==4.4.2
distlib==0.3.3
exceptiongroup==1.1.1
filelock==3.0.12
future==0.18.3
huepy==1.2.1
idna==3.3
imageio==2.26.0
imageio-ffmpeg==0.4.8
iniconfig==2.0.0
lxml==4.7.1
mock==5.0.1
moviepy==1.0.3
multitasking==0.0.10
numpy==1.22.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
platformdirs==2.3.0
pluggy==1.0.0
proglog==0.1.10
pycodestyle==2.10.0
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.10.1
responses==0.23.1
schedule==1.1.0
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
types-PyYAML==6.0.12.8
urllib3==1.26.8
virtualenv==20.8.0
yfinance==0.1.68
python-dotenv~=1.0.0
instagrapi~=1.17.1